Bring json tree getItemString customization from Raw Messages to 3DViz
This includes time formatting, vector lengths, etc. FG-239
